requirements.txt
# --------- pytorch --------- #
torch
datasets
lightning
torchvision
torchaudio
torchtext
torchmetrics
transformers
huggingface_hub
accelerate

# --------- ML --------- #
scipy
opencv-python
einops
matplotlib
scikit-learn
numpy
pandas
seaborn

# --------- Advanced --------- #
# salesforce-lavis
# open_clip_torch
peft
# dask-cuda

# --------- jupyter --------- #
ipykernel       # for jupyter notebook
jupyter         # for jupyter notebook
pyqt5           # for jupyter notebook
pyqtwebengine   # for jupyter notebook

# --------- configs ---------#
hydra-core
hydra-colorlog
hydra-optuna-sweeper
omegaconf
wandb

# --------- utils ---------#
tqdm
Pillow

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
sh              # for running bash commands in some tests (linux/macos only)
h5py            # for datasets
lz4             # for datasets
filelock        # for datasets
geoopt          # for Riemannian optimization
auto_mix_prep   # for data preparation
pipreqs        # for generating requirements.txt


# --------- URL --------- #
# pip install \
#     --extra-index-url=https://pypi.nvidia.com \
#     "cudf-cu12==25.8.*" "dask-cudf-cu12==25.8.*" "cuml-cu12==25.8.*" \
#     "cugraph-cu12==25.8.*" "nx-cugraph-cu12==25.8.*" "cuxfilter-cu12==25.8.*" \
#     "cucim-cu12==25.8.*" "pylibraft-cu12==25.8.*" "raft-dask-cu12==25.8.*" \
#     "cuvs-cu12==25.8.*" "nx-cugraph-cu12==25.8.*"